Altered and added types (MollomError, MollonResponse).
The MollomError type now also has a type specifying that an
HTTP connection error occured, with the connection wrapped in a
CMollomError constructor.

The MollomResponse type was added to contain details about the
response received from the HTTP server, namely the response code,
its accompanying message and the payload of the response. For now the
latter is a String, this should be parsed into a set of well-typed
Haskell values.